Remove files from core which were moved to the RoddyToolLib
