+++ .idea/modules.xml
@@ -5,6 +5,7 @@
       <module fileurl="file://$PROJECT_DIR$/dist/plugins/DefaultPlugin/DefaultPlugin.iml" filepath="$PROJECT_DIR$/dist/plugins/DefaultPlugin/DefaultPlugin.iml" />
       <module fileurl="file://$PROJECT_DIR$/dist/plugins/PluginBase/PluginBase.iml" filepath="$PROJECT_DIR$/dist/plugins/PluginBase/PluginBase.iml" />
       <module fileurl="file://$PROJECT_DIR$/RoddyCore/RoddyCore.iml" filepath="$PROJECT_DIR$/RoddyCore/RoddyCore.iml" />
+      <module fileurl="file://$PROJECT_DIR$/../RoddyInfra/RoddyToolLib/RoddyToolLib.iml" filepath="$PROJECT_DIR$/../RoddyInfra/RoddyToolLib/RoddyToolLib.iml" />
       <module fileurl="file://$PROJECT_DIR$/dist/plugins/TestPluginWithJarFile/TestPluginWithJarFile.iml" filepath="$PROJECT_DIR$/dist/plugins/TestPluginWithJarFile/TestPluginWithJarFile.iml" />
       <module fileurl="file://$PROJECT_DIR$/dist/plugins/TestPluginWithoutJarFile/TestPluginWithoutJarFile.iml" filepath="$PROJECT_DIR$/dist/plugins/TestPluginWithoutJarFile/TestPluginWithoutJarFile.iml" />
     </modules>
