--- RoddyCore/src/de/dkfz/roddy/StringConstants.java
@@ -1,61 +1,0 @@
-/*
- * Copyright (c) 2016 eilslabs.
- *
- * Distributed under the MIT License (license terms are at https://www.github.com/eilslabs/Roddy/LICENSE.txt).
- */
-
-package de.dkfz.roddy;
-
-/**
- */
-public class StringConstants {
-
-    public static final String SPLIT_PERCENT = "%";
-    public static final String SPLIT_SLASH = "[/]";
-    public static final String SPLIT_BACKSLASH = "[\\]";
-    public static final String SPLIT_HASH = "[#]";
-    public static final String SPLIT_COMMA = "[,]";
-    public static final String SPLIT_STOP = "[.]";
-    public static final String SPLIT_DOUBLESTOP = "[.][.]";
-    public static final String SPLIT_SEMICOLON = "[;]";
-    public static final String SPLIT_COLON = "[:]";
-    public static final String SPLIT_PLUS = "[+]";
-    public static final String SPLIT_MINUS = "[-]";
-    public static final String SPLIT_UNDERSCORE = "[_]";
-    public static final String SPLIT_DOUBLEQUOTE = "[\"]";
-    public static final String SPLIT_WHITESPACE = "[ ]";
-    public static final String SPLIT_AT = "[@]";
-    public static final String SPLIT_EQUALS = "[=]";
-    public static final String SPLIT_DOLLAR = "\\$";
-    public static final String SPLIT_DOLLAR_GROOVY = "\\\\\\$";
-    public static final String SPLIT_BRACE_LEFT = "[{]";
-    public static final String SPLIT_BRACE_RIGHT = "[}]";
-    public static final String SPLIT_SBRACKET_LEFT = "\\[";
-    public static final String SPLIT_SBRACKET_RIGHT = "\\[";
-
-    public static final String STOP = ".";
-    public static final String COLON = ":";
-    public static final String SEMICOLON = ";";
-    public static final String COMMA = ",";
-    public static final String HASH = "#";
-    public static final String MINUS = "-";
-    public static final String UNDERSCORE = "_";
-    public static final String PLUS = "+";
-    public static final String TILDE = "~";
-    public static final String DOUBLESTOP = ".";
-    public static final String SBRACKET_LEFT = "[";
-    public static final String SBRACKET_RIGHT = "]";
-    public static final String BRACE_LEFT = "{";
-    public static final String BRACE_RIGHT = "}";
-    public static final String EQUALS = "=";
-    public static final String EMPTY = "";
-    public static final String WHITESPACE = " ";
-    public static final String SINGLE_QUOTE = "'";
-    public static final String DOUBLE_QUOTE = "\"";
-    public static final String FALSE = "false";
-
-    public static final String DOLLAR_LEFTBRACE = "${";
-
-    public static final String TRUE = "true";
-
-}
