+++ RoddyCore/src/de/dkfz/roddy/client/fxuiclient/RoddyUIController.java
@@ -12,7 +12,9 @@ import de.dkfz.roddy.StringConstants;
 import de.dkfz.roddy.client.fxuiclient.settingsviewer.SettingsViewer;
 import de.dkfz.roddy.client.rmiclient.RoddyRMIClientConnection;
 import de.dkfz.roddy.client.rmiclient.RoddyRMIInterfaceImplementation;
-import de.dkfz.roddy.config.AppConfig;
+import de.dkfz.roddy.tools.RoddyConversionHelperMethods;
+import de.dkfz.roddy.tools.RoddyIOHelperMethods;
+import de.dkfz.roddy.tools.AppConfig;
 import de.dkfz.roddy.config.ConfigurationFactory;
 import de.dkfz.roddy.config.InformationalConfigurationContent;
 import de.dkfz.roddy.core.*;
@@ -22,9 +24,6 @@ import de.dkfz.roddy.client.fxuiclient.fxdatawrappers.FXICCWrapper;
 import de.dkfz.roddy.client.fxuiclient.fxwrappercontrols.GenericListViewItemCellImplementation;
 import de.dkfz.roddy.client.fxuiclient.fxwrappercontrols.ProjectTreeItemCellImplementation;
 import de.dkfz.roddy.client.fxuiclient.fxwrappercontrols.ProjectTreeItemCellListener;
-import de.dkfz.roddy.execution.jobs.JobState;
-import de.dkfz.roddy.tools.RoddyConversionHelperMethods;
-import de.dkfz.roddy.tools.RoddyIOHelperMethods;
 import javafx.application.Application;
 import javafx.application.Platform;
 import javafx.beans.value.ChangeListener;
