+++ RoddyCore/src/de/dkfz/roddy/config/ConfigurationFactory.groovy
@@ -24,6 +24,9 @@ import de.dkfz.roddy.core.*
 
 //import de.dkfz.roddy.knowledge.files.*
 import de.dkfz.roddy.plugins.*
+import de.dkfz.roddy.tools.BufferValue
+import de.dkfz.roddy.tools.TimeUnit
+import de.dkfz.roddy.tools.RoddyIOHelperMethods
 import groovy.transform.TypeCheckingMode
 import groovy.util.slurpersupport.*
 import org.apache.commons.io.filefilter.WildcardFileFilter
