+++ RoddyCore/src/de/dkfz/roddy/config/validation/XSDValidator.groovy
@@ -11,15 +11,13 @@ import de.dkfz.roddy.Roddy
 import de.dkfz.roddy.StringConstants;
 import de.dkfz.roddy.config.Configuration
 import de.dkfz.roddy.config.ConfigurationFactory;
-import de.dkfz.roddy.config.InformationalConfigurationContent;
+import de.dkfz.roddy.config.InformationalConfigurationContent
 import de.dkfz.roddy.tools.RoddyIOHelperMethods
 import org.xml.sax.ErrorHandler
 
 import javax.xml.XMLConstants
 import javax.xml.transform.stream.StreamSource
-import javax.xml.validation.SchemaFactory;
-import java.io.File
-import java.util.logging.Logger;
+import javax.xml.validation.SchemaFactory
 
 /**
  * Created by michael on 30.04.15.
