+++ RoddyCore/src/de/dkfz/roddy/execution/io/LocalExecutionService.groovy
@@ -15,7 +15,6 @@ import de.dkfz.roddy.execution.jobs.JobManager
 import de.dkfz.roddy.tools.LoggerWrapper
 
 import java.lang.reflect.Field
-import java.util.logging.Logger
 
 /**
  * The local execution service executes commands on the local machine. For this groovy's execute() method is used.
