+++ RoddyCore/src/de/dkfz/roddy/execution/io/fs/BashCommandSet.groovy
@@ -8,8 +8,6 @@ package de.dkfz.roddy.execution.io.fs
 
 import de.dkfz.roddy.config.converters.BashConverter
 import de.dkfz.roddy.config.converters.ConfigurationConverter
-import de.dkfz.roddy.execution.io.ExecutionService
-import de.dkfz.roddy.tools.RoddyIOHelperMethods
 
 /**
  * Provides a command generator for linux file systems / bash
