+++ RoddyCore/src/de/dkfz/roddy/execution/jobs/cluster/pbs/PBSJobManager.groovy
@@ -13,12 +13,13 @@ import de.dkfz.roddy.tools.BufferUnit;
 import de.dkfz.roddy.execution.io.ExecutionService;
 import de.dkfz.roddy.execution.io.fs.FileSystemAccessProvider
 import de.dkfz.roddy.execution.jobs.cluster.ClusterJobManager;
-import de.dkfz.roddy.knowledge.nativeworkflows.GenericJobInfo;
-import de.dkfz.roddy.tools.*;
+import de.dkfz.roddy.knowledge.nativeworkflows.GenericJobInfo
 import de.dkfz.roddy.config.*;
 import de.dkfz.roddy.core.ExecutionContext;
 import de.dkfz.roddy.execution.io.ExecutionResult;
 import de.dkfz.roddy.execution.jobs.*
+import de.dkfz.roddy.tools.RoddyConversionHelperMethods
+import de.dkfz.roddy.tools.RoddyIOHelperMethods
 
 import java.util.*;
 import java.util.concurrent.locks.ReentrantLock;
