+++ RoddyCore/src/de/dkfz/roddy/plugins/LibrariesFactory.groovy
@@ -17,6 +17,7 @@ import de.dkfz.roddy.tools.*
 import de.dkfz.roddy.tools.Tuple2
 import de.dkfz.roddy.StringConstants
 import de.dkfz.roddy.core.Initializable
+import de.dkfz.roddy.tools.RoddyIOHelperMethods
 import groovy.transform.TypeCheckingMode
 
 import java.lang.reflect.Method
